requirements.txt
fastapi>=0.110
uvicorn>=0.29
pydantic>=2.6
python-dotenv>=1.0
transformers>=4.40
torch>=2.2
sqlalchemy>=2.0
psycopg2-binary>=2.9
sentencepiece